+++ Changelog.md
@@ -3,6 +3,7 @@
 ##Version 1.1.1
 **IN PROGRESS**
 + added LogLevel setting
+* all settings not yet set are now assigned default value on startup
 * generalised "Experimental" settings section to "Advanced" section, with a special prefix for experimental settings
 * fixed bug causing player-size and bg-url input boxes to be falsely disabled